+++ run_integration_tests.sh
@@ -6,4 +6,4 @@ set -eu -o pipefail
 cabal build
 NEWPYTHONPATH="$(pwd)/integration_tests:$PYTHONPATH"
 export PYTHONPATH="${NEWPYTHONPATH}"
-exec timeout 60s python3 integration_tests/orchestrate_integration_tests.py
+exec timeout 90s python3 integration_tests/orchestrate_integration_tests.py